backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
rank-bm25==0.2.2
numpy==2.2.1
requests==2.32.3
PyPDF2==3.0.1
python-docx==1.1.2
Pillow==11.1.0
pytesseract==0.3.13
transformers==4.47.1
torch==2.5.1
